Move UserPermissions component to dashboard
