+++ src/modules/dashboard/components/UserPermissions/UserPermissions.css.d.ts
