+++ src/modules/dashboard/components/UserPermissions/index.ts
@@ -1,0 +1,1 @@
+export { default } from './UserPermissions';
